requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
boto3==1.16.51
botocore==1.19.51
click==7.1.2
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.5.0
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
